requests.elapsed.microsends is not what one think
requests.elapsed.microsends is a completely misleading thing and gives
only microsends part of the overall duration. It is split in seconds and
microseconds. Instead a total_seconds * 1000 should be used (same as in
OpenStackSDK). On most platforms total_seconds is a float giving enough
precision to get milliseconds from it.